Add EnvProcess and logo to UI
