+++ gateway/requests/requests.go
@@ -75,4 +75,5 @@ type Function struct {
 	Image           string  `json:"image"`
 	InvocationCount float64 `json:"invocationCount"` // TODO: shouldn't this be int64?
 	Replicas        uint64  `json:"replicas"`
+	EnvProcess      string  `json:"envProcess"`
 }
